stm32/rtc: cleanup and consolidate
