+++ examples/stm32f4/Cargo.toml
@@ -26,6 +26,7 @@ nb = "1.0.0"
 embedded-storage = "0.3.0"
 micromath = "2.0.0"
 static_cell = "1.0"
+chrono = { version = "^0.4", default-features = false}
 
 [profile.release]
 debug = 2
